rc4 version support compatibility

+++ bcos-framework/interfaces/consensus/ConsensusInterface.h
@@ -66,7 +66,7 @@ public:
     virtual void notifyHighestSyncingNumber(bcos::protocol::BlockNumber _number) = 0;
 
     virtual void asyncNoteUnSealedTxsSize(
-        size_t _unsealedTxsSize, std::function<void(Error::Ptr)> _onRecvResponse) = 0;
+        uint64_t _unsealedTxsSize, std::function<void(Error::Ptr)> _onRecvResponse) = 0;
 
     // get the consensusNodeList
     // Note: if separate sealer with the PBFT module, should implement with notify
