+++ bcos-framework/interfaces/txpool/TxPoolInterface.h
@@ -56,7 +56,7 @@ public:
      * @param _avoidTxs list of transactions that need to be filtered
      * @param _sealCallback after the  txpool responds to the sealed txs, the callback is triggered
      */
-    virtual void asyncSealTxs(size_t _txsLimit, TxsHashSetPtr _avoidTxs,
+    virtual void asyncSealTxs(uint64_t _txsLimit, TxsHashSetPtr _avoidTxs,
         std::function<void(Error::Ptr, bcos::protocol::Block::Ptr, bcos::protocol::Block::Ptr)>
             _sealCallback) = 0;
 
@@ -108,7 +108,7 @@ public:
 
     // for RPC to get pending transactions
     virtual void asyncGetPendingTransactionSize(
-        std::function<void(Error::Ptr, size_t)> _onGetTxsSize) = 0;
+        std::function<void(Error::Ptr, uint64_t)> _onGetTxsSize) = 0;
 
     // notify to reset the txpool when the consensus module startup
     virtual void asyncResetTxPool(std::function<void(Error::Ptr)> _onRecvResponse) = 0;
